refactor delivery module and tracker
