+++ aicostmanager/tracker.py
@@ -8,12 +8,13 @@ from typing import Any, Dict, Optional
 from uuid import uuid4
 
 import httpx
+import logging
 
-from .delivery_manager import (
-    DeliveryManagerType,
-    ImmediateDeliveryManager,
-    MemQueueDeliveryManager,
-    PersistentQueueDeliveryManager,
+from .delivery import (
+    DeliveryType,
+    ImmediateDelivery,
+    MemQueueDelivery,
+    PersistentDelivery,
 )
 from .ini_manager import IniManager
 
@@ -24,7 +25,7 @@ class Tracker:
     def __init__(
         self,
         *,
-        delivery_type: DeliveryManagerType | None = None,
+        delivery_type: DeliveryType | None = None,
         aicm_api_key: Optional[str] = None,
         aicm_api_base: Optional[str] = None,
         aicm_api_url: Optional[str] = None,
@@ -42,6 +43,15 @@ class Tracker:
         transport: httpx.BaseTransport | None = None,
         log_bodies: bool = False,
     ) -> None:
+        log_file = log_file or os.getenv("AICM_TRACKER_LOG_FILE")
+        log_level = (log_level or os.getenv("AICM_TRACKER_LOG_LEVEL", "INFO")).upper()
+        self.logger = logging.getLogger(__name__)
+        self.logger.setLevel(getattr(logging, log_level, logging.INFO))
+        if not self.logger.handlers:
+            handler = logging.FileHandler(log_file) if log_file else logging.StreamHandler()
+            formatter = logging.Formatter("%(asctime)s %(levelname)s %(message)s")
+            handler.setFormatter(formatter)
+            self.logger.addHandler(handler)
         ini_path = (
             aicm_ini_path
             or os.getenv("AICM_INI_PATH")
@@ -50,19 +60,21 @@ class Tracker:
         self.ini_manager = IniManager(ini_path)
         if delivery_type is None:
             name = self.ini_manager.get_option(
-                "tracker", "delivery_manager", DeliveryManagerType.IMMEDIATE.value
+                "tracker", "delivery_manager", DeliveryType.IMMEDIATE.value
             )
-            delivery_type = DeliveryManagerType(name)
-        if delivery_type is DeliveryManagerType.IMMEDIATE:
-            self.delivery = ImmediateDeliveryManager(
+            delivery_type = DeliveryType(name)
+        if delivery_type is DeliveryType.IMMEDIATE:
+            self.delivery = ImmediateDelivery(
                 aicm_api_key=aicm_api_key,
                 aicm_api_base=aicm_api_base,
                 aicm_api_url=aicm_api_url,
                 timeout=timeout,
                 transport=transport,
+                log_file=log_file,
+                log_level=log_level,
             )
-        elif delivery_type is DeliveryManagerType.MEM_QUEUE:
-            self.delivery = MemQueueDeliveryManager(
+        elif delivery_type is DeliveryType.MEM_QUEUE:
+            self.delivery = MemQueueDelivery(
                 aicm_api_key=aicm_api_key,
                 aicm_api_base=aicm_api_base,
                 aicm_api_url=aicm_api_url,
@@ -72,9 +84,11 @@ class Tracker:
                 batch_interval=batch_interval,
                 max_batch_size=max_batch_size,
                 transport=transport,
+                log_file=log_file,
+                log_level=log_level,
             )
         else:
-            self.delivery = PersistentQueueDeliveryManager(
+            self.delivery = PersistentDelivery(
                 aicm_api_key=aicm_api_key,
                 aicm_api_base=aicm_api_base,
                 aicm_api_url=aicm_api_url,
